move unit to sql package
