+++ shardingsphere-infra/shardingsphere-infra-executor/src/main/java/org/apache/shardingsphere/infra/executor/sql/execute/raw/callback/impl/RawSQLExecuteQueryExecutorCallback.java
@@ -18,7 +18,7 @@
 package org.apache.shardingsphere.infra.executor.sql.execute.raw.callback.impl;
 
 import org.apache.shardingsphere.infra.executor.sql.QueryResult;
-import org.apache.shardingsphere.infra.executor.unit.RawSQLExecuteUnit;
+import org.apache.shardingsphere.infra.executor.sql.unit.RawSQLExecuteUnit;
 import org.apache.shardingsphere.infra.executor.sql.execute.raw.callback.RawSQLExecutorCallback;
 
 import java.util.Collection;
